Add performance test for SubstitutionHashEscape and make target
- Add performance-only test file internal/filevalidator/encoding/substitution_hash_escape_performance_test.go using build tag performance.
- Move/remove the existing TestSubstitutionHashEscape_PerformanceRegression from the regular tests to the performance-only test to avoid slowing unit test runs.
- Update Makefile to include a performance-test target that runs the tests with -tags performance.
- Establish baseline thresholds for ns/op and bytes allocated per operation for several encoder scenarios to detect regressions.
